% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{aic}
\alias{aic}
\title{AIC}
\usage{
aic(m = m, w = NULL, ...)
}
\arguments{
\item{m}{glm or lm model}

\item{w}{Weights, used when boot_method="blb"}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
AIC
}
