% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{brier_score}
\alias{brier_score}
\title{Brier score}
\usage{
brier_score(p, y, w, ...)
}
\arguments{
\item{p}{Probabilities from glm}

\item{y}{Outcome variable}

\item{w}{Weights, used when boot_method="blb"}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
Brier score
}
