% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_pgsmetrics.R
\name{effects_pgsmetrics}
\alias{effects_pgsmetrics}
\title{Create coefficient table containing PGS effect size adjusted for covariates.}
\usage{
effects_pgsmetrics(
  data,
  pgs,
  dep,
  covars = NULL,
  missing = "warn",
  report_covars = FALSE
)
}
\arguments{
\item{data}{A data.table containing necessary data: PGS, dependent variable, and covariates.}

\item{pgs}{Character vector. Metrics are computed for these pgs. Must exist in \code{data}.}

\item{dep}{Character. Name of the dependent variable in \code{data}. Binary variable must be 0/1 (otherwise continuous assumed).}

\item{covars}{Character vector. Names of covariate to adjust for. Must be in columns in \code{data}.}

\item{missing}{Warns if missing values in \code{data}. Use missing="drop" to drop rows with missing values. However, we recommend imputing missing values instead. Default: "warn".}

\item{report_covars}{Logical. Include all covariate effects in returned data.frame}
}
\value{
A data.frame with calculated covariate-adjusted effect size of PGS.
\item{coef_table}{A data.frame with calculated covariate-adjusted effect size of PGS.}
}
\description{
Create coefficient table containing PGS effect size adjusted for covariates.
}
\examples{
# Simulate data
data <- simulate_data(n = 1000, n_pgs = 3)

# raw PGS coefficients 
coef_table <- effects_pgsmetrics(data,
    pgs = c("pgs1", "pgs2", "pgs3"),
    dep = "y_bin",
    covars = c("age", "sex")
)
# print(coef_table)

# Obtain coefficients for top decile of PGS
data$pgs1.top10 <- 1*(data$pgs1 >= quantile(data$pgs1, 0.9, na.rm = TRUE))

coef <- effects_pgsmetrics(data, 
    pgs = c("pgs1", "pgs1.top10"), 
    dep = "y_bin", 
    covars = c("age", "sex"),
 )
# print(coef)

}
