% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_metrics.R
\name{filter_metrics}
\alias{filter_metrics}
\title{Filter a list of metrics by type}
\usage{
filter_metrics(metrics_list, model_type)
}
\arguments{
\item{metrics_list}{List of metrics}

\item{model_type}{binary/continuous/both.}
}
\value{
A list of all registered metrics.
}
\description{
Filter a list of metrics by type
}
