% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit Generalized Linear Model}
\usage{
fit_glm(x, y, family = binomial(), weights = NULL, fastglm_method = 2)
}
\arguments{
\item{x}{Matrix of predictors}

\item{y}{Vector of responses}
}
\value{
Fitted GLM object
}
\description{
Fit Generalized Linear Model
}
\keyword{internal}
