% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(p = p, y = y, w = w, ...)
}
\arguments{
\item{p}{Predictions from lm}

\item{y}{Outcome vector}

\item{w}{Weights, used when boot_method="blb"}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
Mean Absolute Error
}
