% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_metrics.R
\name{Metric}
\alias{Metric}
\title{Metric Class}
\description{
An R6 class representing a metric for evaluating Polygenic Scores (PGS).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Character. The name of the metric.}

\item{\code{fn}}{Function. The function used to calculate the metric.}

\item{\code{type}}{binary/continuous/both.}

\item{\code{description}}{Longer name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metric-new}{\code{Metric$new()}}
\item \href{#method-Metric-calculate}{\code{Metric$calculate()}}
\item \href{#method-Metric-clone}{\code{Metric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-new"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-new}{}}}
\subsection{Method \code{new()}}{
Create a new Metric object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$new(name, fn, type, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. The name of the metric.}

\item{\code{fn}}{Function. The function used to calculate the metric.}

\item{\code{type}}{binary/continuous/both.}

\item{\code{description}}{Longer name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate the metric value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$calculate(
  m = NULL,
  m_lm = NULL,
  p = NULL,
  y = NULL,
  liab_prep = NULL,
  n = NULL,
  w = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{glm model}

\item{\code{m_lm}}{lm model (family='gaussian')}

\item{\code{p}}{probabilities from glm}

\item{\code{y}}{response}

\item{\code{liab_prep}}{data input to r2_liability}

\item{\code{n}}{number of observations}

\item{\code{w}}{weights}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The calculated metric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
