% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgsmetrics.R
\name{pgsmetrics}
\alias{pgsmetrics}
\title{Compare Polygenic Scores (PGS) using bootstrapping}
\usage{
pgsmetrics(
  data,
  pgs,
  dep,
  covars = NULL,
  K = NULL,
  boot = TRUE,
  boot_method = "standard",
  blb_s = 10,
  blb_r = 50,
  blb_b = 0.9,
  n_boot = 1000,
  n_cores = 1,
  missing = "warn",
  return_boot_stats = FALSE,
  custom_metrics_list = NULL
)
}
\arguments{
\item{data}{A data.table containing necessary data: PGS, dependent variable, and covariates.}

\item{pgs}{Character vector. Metrics are computed for these pgs. Must exist in \code{data}.}

\item{dep}{Character. Name of the dependent variable in \code{data}. Binary variable must be 0/1 (otherwise continuous assumed).}

\item{covars}{Character vector. Names of covariate to adjust for. Must be in columns in \code{data}.}

\item{K}{Double, between 0 and 1. Prevalence of the binary trait in the population. Used by $R^2$ on the liability scale.}

\item{boot}{Boolean. Whether to perform bootstrap or not.}

\item{boot_method}{Character vector. Either \code{standard} or \code{blb}. \code{standard} is the usual non-parametric bootstrap, \code{blb} is the Bag of Little Bootstraps, useful for analysis on large datasets (f.x., $n>500000$) with many PGS. However, BLB and the standard bootstrap may not yield identical results. The standard bootstrap uses the observed statistics (all observations) and gives percentile CIs with resampling. The BLB bootstrap estimates the mean across all resamples (instead of the observed value) and percentile CIs.}

\item{blb_s}{Integer. TODO. Default: 20.}

\item{blb_r}{Integer. TODO. Default: 100.}

\item{blb_b}{Double between 0 and 1. TODO. Default: 0.8.}

\item{n_boot}{Integer. Number of bootstrap iterations. Default is 1000.}

\item{n_cores}{Integer. Number of cores for parallel processing. Default is 1.}

\item{missing}{Warns if missing values in \code{data}. Use missing="drop" to drop rows with missing values. However, we recommend imputing missing values instead. Default: "warn".}

\item{return_boot_stats}{Return bootstrap output before processing (for debugging). Default: FALSE.}

\item{custom_metrics_list}{List of custom metrics to be summarized.}
}
\value{
An object of class "pgsmetrics" containing:
\item{metrics}{A data.table with calculated metrics for each PGS. In this data.table, \code{type} can be partial/full/cov, where partial=full-cov, full=model with covariates+pgs, cov=model with only covariates. When using \code{boot_method="blb"}, the \code{observed} column is the median value of the metric statistic, but when using \code{boot_method="standard"}, it is the value from an analysis on the unmodified dataset.}
\item{boot_output}{The raw output from the bootstrapping process.}
\item{call}{The original function call.}
}
\description{
This function compares Polygenic Scores (PGS) by bootstrapping the difference
between them using various metrics. It supports single PGS evaluation, pairwise
comparisons, and ranking of multiple PGS.
}
\examples{
# Simulate data
data <- simulate_data(n = 1000, n_pgs = 3)

# Single PGS evaluation
result_single <- pgsmetrics(data,
    pgs = c("pgs1", "pgs2", "pgs3"),
    dep = "y_bin",
    covars = c("age", "sex"),
    n_boot = 100
)
# print(result_single)

}
