% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_pgsmetrics.R
\name{plot.pgsmetrics}
\alias{plot.pgsmetrics}
\title{Plot method for pgsmetrics objects}
\usage{
\method{plot}{pgsmetrics}(x, type = "metrics", ...)
}
\arguments{
\item{x}{An object of class "pgsmetrics".}

\item{type}{What results to plot? Options: metrics/ranks/diff.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Plot method for pgsmetrics objects
}
