% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot Calibration Curve with Prediction Intervals}
\usage{
plot_calibration(
  df,
  pgs,
  dep,
  covars = NULL,
  x_lab = "Predicted",
  y_lab = "Observed",
  conf_level = 0.95,
  annotation_size = 3,
  annotation_x = 0.05,
  annotation_y = 0.85
)
}
\arguments{
\item{df}{A data.frame object, must contain pgs and dep.}

\item{pgs}{PGS to include in model.}

\item{dep}{Outcome variable, must be binary.}

\item{covars}{Additional covariates to include in the model. Default: NULL.}

\item{x_lab}{Text, x-axis. Default: "Predicted".}

\item{y_lab}{Text, y-axis. Default: "Observed".}

\item{conf_level}{Confidence level for prediction interval. Default: 0.95.}

\item{annotation_size}{Font size in  calibration metric text box. Default: 3.}

\item{annotation_x}{x-axis position of the calibration metric text box. Default: 0.05.}

\item{annotation_y}{y-axis position of the calibration metric text box. Default: 0.85.}
}
\value{
A ggplot object with the following annotated calibration metrics:
\item{ICI}{Integrated calibration index (average absolute difference)}
\item{E50}{Median absolute difference}
\item{E90}{90th percentile absolute difference}
\item{Emax}{Maximum absolute difference}
\item{ECI}{Estimated calibration index (average square difference)}
}
\description{
Plot Calibration Curve with Prediction Intervals
}
