% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dcurve.R
\name{plot_dcurve}
\alias{plot_dcurve}
\title{Plot Decision Curve}
\usage{
plot_dcurve(
  df,
  pgs,
  dep,
  covars = NULL,
  dca_thresholds = seq(0, 0.5, by = 0.01),
  dca_plot_smooth = TRUE
)
}
\arguments{
\item{df}{A data.frame object, must contain pgs and dep.}

\item{pgs}{PGS to include in model.}

\item{dep}{Outcome variable, must be binary.}

\item{covars}{Additional covariates to include in the model. Default: NULL.}

\item{dca_thresholds}{Vector of threshold probabilities, x-axis ('thresholds' argument in dcurves::dca.)}

\item{dca_plot_smooth}{Smooth lines? ('smooth' argument in dcurves:::plot.dca.)}
}
\value{
A ggplot object.
}
\description{
Plot Decision Curve
}
