% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot PGS Metrics}
\usage{
plot_metrics(
  pgsmetrics_object,
  type = "metrics",
  plot_type = "points",
  model_type = "partial",
  order_by_metric = "aic",
  title = NULL,
  nrow = NULL,
  ncol = 1
)
}
\arguments{
\item{pgsmetrics_object}{TODO.}

\item{type}{TODO.}

\item{plot_type}{TODO.}

\item{model_type}{TODO.}

\item{order_by_metric}{TODO.}

\item{title}{TODO.}

\item{nrow}{TODO.}

\item{ncol}{TODO.}
}
\value{
A ggplot object
}
\description{
Plot PGS Metrics
}
