% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_pgsmetrics.R
\name{print.pgsmetrics}
\alias{print.pgsmetrics}
\title{Print method for pgsmetrics objects}
\usage{
\method{print}{pgsmetrics}(x, n = 10, ...)
}
\arguments{
\item{x}{An object of class "pgsmetrics".}

\item{n}{Number of rows to print.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for pgsmetrics objects
}
