% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2}
\alias{r2}
\title{$R^2$}
\usage{
r2(m = m, w = w, ...)
}
\arguments{
\item{m}{lm model}

\item{w}{Weights, used when boot_method="blb"}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
$R^2$
}
