% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2_liability}
\alias{r2_liability}
\title{$R^2$ on the liability scale}
\usage{
r2_liability(m_lm = m_lm, liab_prep = liab_prep, w = w, ...)
}
\arguments{
\item{m_lm}{lm model (family="gaussian")}

\item{liab_prep}{Data from \code{r2_liability_prep}}

\item{w}{Weights, used when boot_method="blb"}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
$R^2$ on the liability scale
}
