% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2_liability_prep}
\alias{r2_liability_prep}
\title{Prep data for \code{r2_liability}}
\usage{
r2_liability_prep(dep, K = NULL, w = NULL)
}
\arguments{
\item{dep}{Outcome vector}

\item{K}{population prevalence, if NULL then =P}

\item{w}{Weights, used when boot_method="blb"}
}
\description{
Prep data for \code{r2_liability}
}
\keyword{internal}
