% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2_nagelkerke}
\alias{r2_nagelkerke}
\title{Nagelkerke $R^2$}
\usage{
r2_nagelkerke(m = m, n = n, ...)
}
\arguments{
\item{m}{glm or lm model (family="binomial" or family="gaussian")}

\item{n}{Total number of observations}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
Nagelkerke $R^2$
}
