% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_metrics.R
\name{register_metric}
\alias{register_metric}
\title{Metric Registry
Register a new metric}
\usage{
register_metric(metrics_list, name, fn, type, description)
}
\arguments{
\item{metrics_list}{List of metrics}

\item{name}{Character. Name of the metric.}

\item{fn}{Function. The metric calculation function.}

\item{type}{binary/continuous/both.}

\item{description}{Description.}
}
\description{
Metric Registry
Register a new metric
}
