% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{roc_auc}
\alias{roc_auc}
\title{ROC AUC}
\usage{
roc_auc(p = p, y = y, w = w, ...)
}
\arguments{
\item{p}{Probabilities from glm}

\item{y}{Outcome vector}

\item{w}{Weights, used when boot_method="blb"}

\item{...}{Additional arguments}
}
\value{
Numeric
}
\description{
ROC AUC
}
