% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data for testing}
\usage{
simulate_data(
  n = 10000,
  seed = NULL,
  n_pgs = 5,
  pgs_true_her_explained = 0.25,
  pgs_her_explained_ub = pgs_true_her_explained/2,
  pgs_her_explained_lb = pgs_true_her_explained/4,
  age_prop_explained = 0.25,
  env_prop_explained = 0.5,
  dep_prevalence = 0.25
)
}
\arguments{
\item{n}{The number of observations to simulate. Default: 10000.}

\item{seed}{The random seed for reproducibility. Default: NULL.}

\item{n_pgs}{The number of PGS to simulate. Default: 5.}

\item{pgs_true_her_explained}{Has to be >= 0 and <= 0.5. Default: 0.25.}

\item{pgs_her_explained_ub}{Lower bound of heritability explained by PGS.}

\item{pgs_her_explained_lb}{Upper bound of heritability explained by PGS.}

\item{age_prop_explained}{The proportion of \code{y} and \code{y_bin} explained by age. Default: 0.25.}

\item{env_prop_explained}{The proportion of \code{y} and \code{y_bin} explained by other environmental factors. Default: 0.5.}

\item{dep_prevalence}{Prevalence of the binary trait. Default: 0.25.}
}
\value{
A data.table with simulated data
}
\description{
This function simulates a dataset with PGS, covariates, and dependent variables
for testing purposes.
}
\examples{
# Simulate data
data <- simulate_data(n = 1000, n_pgs = 3)
print(data)

}
