% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_pgsmetrics.R
\name{summary.pgsmetrics}
\alias{summary.pgsmetrics}
\title{Summary method for pgsmetrics objects}
\usage{
\method{summary}{pgsmetrics}(object, ...)
}
\arguments{
\item{object}{An object of class "pgsmetrics".}

\item{...}{Additional arguments passed to summary.}
}
\description{
Summary method for pgsmetrics objects
}
