% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs for pgsmetrics function}
\usage{
validate_inputs(
  data,
  K,
  pgs,
  dep,
  covars,
  boot_method,
  blb_s,
  blb_r,
  blb_b,
  n_boot,
  n_cores,
  boot,
  missing
)
}
\arguments{
\item{data}{A data.table containing necessary data.}

\item{pgs}{Character vector. Names of PGS columns.}

\item{dep}{Character. Name of the dependent variable column.}

\item{covars}{Character vector. Names of covariate columns.}

\item{n_boot}{Integer. Number of bootstrap iterations.}

\item{n_cores}{Integer. Number of cores for parallel processing.}
}
\value{
NULL. Throws an error if any input is invalid.
}
\description{
Validate inputs for pgsmetrics function
}
\keyword{internal}
